refactor(create-webstone-app): rename directories
