+++ packages/create-webstone-app/src/tasks/1-create-directory/index.ts
