+++ packages/create-webstone-app/src/tasks/2-install-dependencies/index.ts
