+++ packages/create-webstone-app/src/tasks/index.ts
@@ -5,8 +5,8 @@ const enquirer = new Enquirer();
 
 import { Ctx } from "../helpers";
 import { tasks as metadataTasks } from "./0-determine-context";
-import { tasks as createAppDirectoryTasks } from "./1-create-app-directory/index";
-import { tasks as installAppDependenciesTasks } from "./2-install-app-dependencies/index";
+import { tasks as createAppDirectoryTasks } from "./1-create-directory/index";
+import { tasks as installAppDependenciesTasks } from "./2-install-dependencies/index";
 import { tasks as configureApp } from "./3-configure-app";
 
 export const tasks = new Listr<Ctx>(
