+++ packages/create-webstone-app/template/plugin/tsconfig.json
@@ -1,0 +1,18 @@
+{
+	"compilerOptions": {
+		"module": "CommonJS",
+		"target": "ES5",
+		"allowJs": true,
+		"checkJs": true,
+		"esModuleInterop": true,
+		"forceConsistentCasingInFileNames": true,
+		"resolveJsonModule": true,
+		"skipLibCheck": true,
+		"sourceMap": false,
+		"strict": true,
+		"outDir": "./package/cli/",
+		"declaration": false,
+		"noImplicitAny": false
+	},
+	"include": ["src/lib/cli/**/*"]
+}
